# docs/api/openapi-etapa1.yaml
openapi: 3.1.0
info:
  title: Cerniq.app API — Etapa 1 (Data Enrichment)
  version: 1.0.0
  description: |
    Specificație OpenAPI pentru Etapa 1 - Data Enrichment Pipeline.

    ## Overview
    Etapa 1 implementează pipeline-ul de îmbogățire date pentru prospecți B2B:
    - **Bronze Layer**: Ingestie și validare date brute (CSV, Excel, API, Webhook)
    - **Silver Layer**: Îmbogățire date (ANAF, Termene.ro, ONRC, Geocoding)
    - **Gold Layer**: Lead-uri calificate pentru outreach
    - **HITL**: Human-in-the-Loop pentru aprobare date incerte

    ## Autentificare
    Toate endpoint-urile necesită autentificare JWT:
    ```
    Authorization: Bearer <access_token>
    X-Tenant-ID: <tenant_uuid>
    ```

    ## Rate Limiting
    - Standard: 100 req/min
    - Import: 10 req/min
    - Enrichment: 60 req/min

    Pentru schema globală și convenții comune, vezi docs/api/openapi.yaml.
  contact:
    name: Cerniq Development Team
    email: dev@cerniq.app
  license:
    name: Proprietary
    url: https://cerniq.app/license
servers:
  - url: https://api.cerniq.app/api/v1
    description: Production
  - url: https://staging.cerniq.app/api/v1
    description: Staging
  - url: http://localhost:64000/api/v1
    description: Development
...
paths:
...
  /approvals:
    get:
      tags: [Approvals]
      operationId: listApprovals
      summary: Listează task-uri de aprobare
      description: |
        Listează toate task-urile de aprobare HITL cu paginare și filtrare.
        HITL (Human-in-the-Loop) este folosit pentru validarea datelor incerte.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ApprovalStatus"
        - name: approvalType
          in: query
          schema:
            $ref: "#/components/schemas/ApprovalType"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/ApprovalPriority"
        - name: assignedTo
          in: query
          schema:
            type: string
            format: uuid
        - name: pipelineStage
          in: query
          schema:
            type: string
            enum: [E1, E2, E3, E4, E5]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, dueAt, priority]
            default: dueAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Lista aprobări
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalListResponse"
  /approvals/{approvalId}:
    get:
      tags: [Approvals]
      operationId: getApproval
      summary: Obține detalii aprobare
      description: Returnează detalii complete pentru un task de aprobare
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ApprovalIdParam"
      responses:
        "200":
          description: Detalii aprobare
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /approvals/{approvalId}/assign:
    post:
      tags: [Approvals]
      operationId: assignApproval
      summary: Asignează task de aprobare
      description: Asignează sau de-asignează un task de aprobare unui utilizator
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ApprovalIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assignTo:
                  type: string
                  format: uuid
                  nullable: true
                  description: User ID pentru asignare, null pentru de-asignare
      responses:
        "200":
          description: Task asignat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalAssignResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /approvals/{approvalId}/decide:
    post:
      tags: [Approvals]
      operationId: decideApproval
      summary: Decide pe task de aprobare
      description: |
        Înregistrează decizia (approved/rejected) pentru un task de aprobare.
        Deciziile sunt finale și declanșează acțiunile corespunzătoare în pipeline.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ApprovalIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
              properties:
                decision:
                  type: string
                  enum: [approved, rejected]
                reason:
                  type: string
                  maxLength: 1000
                  description: Motivul deciziei
      responses:
        "200":
          description: Decizie înregistrată
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "400":
          description: Task deja decis sau expirat
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== ENRICHMENT QUEUES ====================
  /enrichment/queues:
    get:
      tags: [Enrichment]
      operationId: listEnrichmentQueues
      summary: Listează queue-uri îmbogățire
      description: Returnează statusul tuturor queue-urilor de îmbogățire BullMQ
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista queue-uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichmentQueuesResponse"
  /enrichment/queues/{queueName}/pause:
    post:
      tags: [Enrichment]
      operationId: pauseEnrichmentQueue
      summary: Pune în pauză queue
      description: Oprește temporar procesarea unui queue de îmbogățire
      security:
        - BearerAuth: []
      parameters:
        - name: queueName
          in: path
          required: true
          schema:
            type: string
          description: Numele queue-ului (e.g., enrich:anaf:fiscal-status)
      responses:
        "200":
          description: Queue în pauză
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStatusResponse"
        "404":
          description: Queue inexistent
  /enrichment/queues/{queueName}/resume:
    post:
      tags: [Enrichment]
      operationId: resumeEnrichmentQueue
      summary: Repornește queue
      description: Repornește procesarea unui queue de îmbogățire
      security:
        - BearerAuth: []
      parameters:
        - name: queueName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue repornit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStatusResponse"
        "404":
          description: Queue inexistent
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obținut din /api/v1/auth/login
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Numărul paginii (1-indexed)
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Număr de rezultate per pagină
...
    ApprovalIdParam:
      name: approvalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID-ul task-ului de aprobare
  responses:
...
    NotFound:
      description: Resursa nu a fost găsită
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    PaginationMeta:
      type: object
      required:
        - total
        - page
        - pageSize
        - totalPages
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer

    # ========== ENUMS ==========
...
    ApprovalStatus:
      type: string
      enum: [pending, assigned, approved, rejected, expired, escalated]
    ApprovalType:
      type: string
      enum: [data_quality, dedup_review, ai_validation]
    ApprovalPriority:
      type: string
      enum: [critical, high, normal, low]

    # ========== DASHBOARD ==========
...
    ApprovalListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalSummary"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    ApprovalSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        entityTitle:
          type: string
        pipelineStage:
          type: string
        approvalType:
          $ref: "#/components/schemas/ApprovalType"
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        priority:
          $ref: "#/components/schemas/ApprovalPriority"
        slaMinutes:
          type: integer
        dueAt:
          type: string
          format: date-time
        assignedTo:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    ApprovalDetailResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          allOf:
            - $ref: "#/components/schemas/ApprovalSummary"
            - type: object
              properties:
                decision:
                  type: string
                  nullable: true
                decisionReason:
                  type: string
                  nullable: true
                decidedBy:
                  type: string
                  format: uuid
                  nullable: true
                decidedAt:
                  type: string
                  format: date-time
                  nullable: true
                assignedAt:
                  type: string
                  format: date-time
                  nullable: true
                entityData:
                  type: object
                  additionalProperties: true
    ApprovalAssignResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              const: assigned
            assignedTo:
              type: string
              format: uuid
            assignedAt:
              type: string
              format: date-time
    ApprovalDecisionResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
            decision:
              type: string
            decisionReason:
              type: string
            decidedBy:
              type: string
              format: uuid
            decidedAt:
              type: string
              format: date-time

    # ========== ENRICHMENT QUEUES ==========
    EnrichmentQueuesResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/QueueStats"
    QueueStats:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        waiting:
          type: integer
        active:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        delayed:
          type: integer
        paused:
          type: boolean
        rateLimit:
          type: object
          properties:
            max:
              type: integer
            duration:
              type: integer
        lastJobAt:
          type: string
          format: date-time
          nullable: true
    QueueStatusResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            name:
              type: string
            paused:
              type: boolean
security:
  - BearerAuth: []